added ticket service and order service

+++ backend/src/main/java/com/csci4050/order/entities/TicketType.java
@@ -1,4 +1,3 @@
-
 package com.csci4050.order.entities;
 
 import jakarta.persistence.Entity;
@@ -11,17 +10,17 @@ public class TicketType {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
+    private Integer id;
 
     private String name;
     private double price;
 
     // Getters and Setters
-    public Long getId() {
+    public Integer getId() {
         return id;
     }
 
-    public void setId(Long id) {
+    public void setId(Integer id) {
         this.id = id;
     }
