+++ backend/src/main/java/com/csci4050/order/repositories/OrderRepository.java
@@ -1,8 +1,8 @@
-
 package com.csci4050.order.repositories;
 
-import com.csci4050.order.entities.Order;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface OrderRepository extends JpaRepository<Order, Long> {
+import com.csci4050.order.entities.Order;
+
+public interface OrderRepository extends JpaRepository<Order, Integer> {
 }
