+++ backend/src/main/java/com/csci4050/order/repositories/TicketTypeRepository.java
@@ -1,7 +1,8 @@
 package com.csci4050.order.repositories;
 
-import com.csci4050.order.entities.TicketType;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface TicketTypeRepository extends JpaRepository<TicketType, Long> {
+import com.csci4050.order.entities.TicketType;
+
+public interface TicketTypeRepository extends JpaRepository<TicketType, Integer> {
 }
